Tweaks iOS changes for better org and use of JSON.
